add support for remote machine
